+++ scopes/workspace/install/exceptions/dependency-type-not-supported-in-policy.ts
