+++ scopes/workspace/install/exceptions/index.ts
