+++ scopes/workspace/install/index.ts
@@ -1,0 +1,5 @@
+import { InstallAspect } from './install.aspect';
+
+export type { InstallMain } from './install.main.runtime';
+export default InstallAspect;
+export { InstallAspect };
