+++ scopes/workspace/install/install.aspect.ts
@@ -1,0 +1,5 @@
+import { Aspect } from '@teambit/harmony';
+
+export const InstallAspect = Aspect.create({
+  id: 'teambit.bit/install',
+});
