+++ scopes/workspace/workspace/workspace.ts
@@ -1,5 +1,4 @@
 /* eslint-disable max-lines */
-import chalk from 'chalk';
 import memoize from 'memoizee';
 import mapSeries from 'p-map-series';
 import type { PubsubMain } from '@teambit/pubsub';
@@ -20,7 +19,6 @@ import {
   InvalidComponent,
   ResolveAspectsOptions,
 } from '@teambit/component';
-import { Importer } from '@teambit/importer';
 import { BitError } from '@teambit/bit-error';
 import { REMOVE_EXTENSION_SPECIAL_SIGN } from '@teambit/legacy/dist/consumer/config';
 import { ComponentScopeDirMap, ConfigMain } from '@teambit/config';
@@ -48,8 +46,6 @@ import { RequireableComponent } from '@teambit/harmony.modules.requireable-compo
 import type { VariantsMain, Patterns } from '@teambit/variants';
 import { link } from '@teambit/legacy/dist/api/consumer';
 import LegacyGraph from '@teambit/legacy/dist/scope/graph/graph';
-import { ImportOptions } from '@teambit/legacy/dist/consumer/component-ops/import-components';
-import { NothingToImport } from '@teambit/legacy/dist/consumer/exceptions';
 import { BitIds } from '@teambit/legacy/dist/bit-id';
 import { BitId, InvalidScopeName, InvalidScopeNameFromRemote, isValidScopeName } from '@teambit/legacy-bit-id';
 import { LaneId } from '@teambit/lane-id';
@@ -68,12 +64,7 @@ import { NoComponentDir } from '@teambit/legacy/dist/consumer/component/exceptio
 import { ExtensionDataList, ExtensionDataEntry } from '@teambit/legacy/dist/consumer/config/extension-data';
 import { pathIsInside } from '@teambit/legacy/dist/utils';
 import componentIdToPackageName from '@teambit/legacy/dist/utils/bit/component-id-to-package-name';
-import {
-  PathOsBased,
-  PathOsBasedRelative,
-  PathOsBasedAbsolute,
-  pathNormalizeToLinux,
-} from '@teambit/legacy/dist/utils/path';
+import { PathOsBased, PathOsBasedRelative, PathOsBasedAbsolute } from '@teambit/legacy/dist/utils/path';
 import fs from 'fs-extra';
 import { slice, uniqBy, difference, compact, pick, partition, isEmpty } from 'lodash';
 import path from 'path';
@@ -88,7 +79,6 @@ import { LaneNotFound } from '@teambit/legacy/dist/api/scope/lib/exceptions/lane
 import { ScopeNotFoundOrDenied } from '@teambit/legacy/dist/remotes/exceptions/scope-not-found-or-denied';
 import { ComponentLoadOptions } from '@teambit/legacy/dist/consumer/component/component-loader';
 import { ComponentConfigFile } from './component-config-file';
-import { DependencyTypeNotSupportedInPolicy } from './exceptions';
 import {
   OnComponentAdd,
   OnComponentChange,
@@ -97,7 +87,6 @@ import {
   OnComponentRemove,
   SerializableResults,
 } from './on-component-events';
-import { pickOutdatedPkgs } from './pick-outdated-pkgs';
 import { WorkspaceExtConfig } from './types';
 import { Watcher, WatchOptions } from './watch/watcher';
 import { ComponentStatus } from './workspace-component/component-status';
@@ -128,22 +117,6 @@ export interface EjectConfOptions {
   propagate?: boolean;
   override?: boolean;
 }
-
-export type WorkspaceInstallOptions = {
-  addMissingPeers?: boolean;
-  variants?: string;
-  lifecycleType?: WorkspaceDependencyLifecycleType;
-  dedupe: boolean;
-  import: boolean;
-  copyPeerToRuntimeOnRoot?: boolean;
-  copyPeerToRuntimeOnComponents?: boolean;
-  updateExisting: boolean;
-  savePrefix?: string;
-};
-
-export type ModulesInstallOptions = Omit<WorkspaceInstallOptions, 'updateExisting' | 'lifecycleType' | 'import'>;
-
-export type WorkspaceLinkOptions = LinkingOptions;
 
 export type TrackData = {
   rootDir: PathOsBasedRelative; // path relative to the workspace
@@ -1770,281 +1743,6 @@ needed-for: ${neededFor || '<unknown>'}`);
     return resolved;
   }
 
-  private async getComponentsDirectory(ids: ComponentID[]): Promise<ComponentMap<string>> {
-    const components = ids.length ? await this.getMany(ids) : await this.list();
-    return ComponentMap.as<string>(components, (component) => this.componentDir(component.id));
-  }
-
-  /**
-   * Install dependencies for all components in the workspace
-   *
-   * @returns
-   * @memberof Workspace
-   */
-  async install(packages?: string[], options?: WorkspaceInstallOptions): Promise<ComponentMap<string>> {
-    if (packages && packages.length) {
-      await this._addPackages(packages, options);
-    }
-    if (options?.addMissingPeers) {
-      const compDirMap = await this.getComponentsDirectory([]);
-      const mergedRootPolicy = this.dependencyResolver.getWorkspacePolicy();
-      const depsFilterFn = await this.generateFilterFnForDepsFromLocalRemote();
-      const pmInstallOptions: PackageManagerInstallOptions = {
-        dedupe: options?.dedupe,
-        copyPeerToRuntimeOnRoot: options?.copyPeerToRuntimeOnRoot ?? true,
-        copyPeerToRuntimeOnComponents: options?.copyPeerToRuntimeOnComponents ?? false,
-        dependencyFilterFn: depsFilterFn,
-        overrides: this.dependencyResolver.config.overrides,
-        packageImportMethod: this.dependencyResolver.config.packageImportMethod,
-      };
-      const missingPeers = await this.dependencyResolver.getMissingPeerDependencies(
-        this.path,
-        mergedRootPolicy,
-        compDirMap,
-        pmInstallOptions
-      );
-      if (missingPeers) {
-        const missingPeerPackages = Object.entries(missingPeers).map(([peerName, range]) => `${peerName}@${range}`);
-        await this._addPackages(missingPeerPackages, options);
-      } else {
-        this.logger.console('No missing peer dependencies found.');
-      }
-    }
-    if (options?.import) {
-      this.logger.setStatusLine('importing missing objects');
-      await this.importObjects();
-      this.logger.consoleSuccess();
-    }
-    return this._installModules(options);
-  }
-
-  private async _addPackages(packages: string[], options?: WorkspaceInstallOptions) {
-    if (!options?.variants && (options?.lifecycleType as string) === 'dev') {
-      throw new DependencyTypeNotSupportedInPolicy(options?.lifecycleType as string);
-    }
-    this.logger.debug(`installing the following packages: ${packages.join()}`);
-    const resolver = await this.dependencyResolver.getVersionResolver();
-    const resolvedPackagesP = packages.map((packageName) =>
-      resolver.resolveRemoteVersion(packageName, {
-        rootDir: this.path,
-      })
-    );
-    const resolvedPackages = await Promise.all(resolvedPackagesP);
-    const newWorkspacePolicyEntries: WorkspacePolicyEntry[] = [];
-    resolvedPackages.forEach((resolvedPackage) => {
-      if (resolvedPackage.version) {
-        const versionWithPrefix = this.dependencyResolver.getVersionWithSavePrefix(
-          resolvedPackage.version,
-          options?.savePrefix
-        );
-        newWorkspacePolicyEntries.push({
-          dependencyId: resolvedPackage.packageName,
-          value: {
-            version: versionWithPrefix,
-          },
-          lifecycleType: options?.lifecycleType || 'runtime',
-        });
-      }
-    });
-    if (!options?.variants) {
-      this.dependencyResolver.addToRootPolicy(newWorkspacePolicyEntries, {
-        updateExisting: options?.updateExisting ?? false,
-      });
-    } else {
-      // TODO: implement
-    }
-    await this.dependencyResolver.persistConfig(this.path);
-  }
-
-  private async _getComponentsWithDependencyPolicies() {
-    const allComponentIds = await this.listIds();
-    const componentConfigFiles: Record<string, ComponentConfigFile> = {};
-    const componentPoliciesById: Record<string, any> = {};
-    (
-      await Promise.all<ComponentConfigFile | undefined>(
-        allComponentIds.map((componentId) => this.componentConfigFile(componentId))
-      )
-    ).forEach((componentConfigFile, index) => {
-      if (!componentConfigFile) return;
-      const depResolverConfig = componentConfigFile.aspects.get(DependencyResolverAspect.id);
-      if (!depResolverConfig) return;
-      const componentId = allComponentIds[index].toString();
-      componentConfigFiles[componentId] = componentConfigFile;
-      componentPoliciesById[componentId] = depResolverConfig.config.policy;
-    });
-    return {
-      componentConfigFiles,
-      componentPoliciesById,
-    };
-  }
-
-  /**
-   * Updates out-of-date dependencies in the workspace.
-   *
-   * @param options.all {Boolean} updates all outdated dependencies without showing a prompt.
-   */
-  async updateDependencies(options: { all: boolean }) {
-    const { componentConfigFiles, componentPoliciesById } = await this._getComponentsWithDependencyPolicies();
-    const variantPatterns = this.variants.raw();
-    const variantPoliciesByPatterns = this._variantPatternsToDepPolicesDict(variantPatterns);
-    const components = await this.list();
-    const outdatedPkgs = await this.dependencyResolver.getOutdatedPkgsFromPolicies({
-      rootDir: this.path,
-      variantPoliciesByPatterns,
-      componentPoliciesById,
-      components,
-    });
-    let outdatedPkgsToUpdate!: OutdatedPkg[];
-    if (options.all) {
-      outdatedPkgsToUpdate = outdatedPkgs;
-    } else {
-      this.logger.off();
-      outdatedPkgsToUpdate = await pickOutdatedPkgs(outdatedPkgs);
-      this.logger.on();
-    }
-    const { updatedVariants, updatedComponents } = this.dependencyResolver.applyUpdates(outdatedPkgsToUpdate, {
-      variantPoliciesByPatterns,
-      componentPoliciesById,
-    });
-    await this._updateVariantsPolicies(variantPatterns, updatedVariants);
-    const updatedComponentConfigFiles = Object.values(pick(componentConfigFiles, updatedComponents));
-    await this._saveManyComponentConfigFiles(updatedComponentConfigFiles);
-    await this._reloadConsumer();
-    return this._installModules({ dedupe: true });
-  }
-
-  private _variantPatternsToDepPolicesDict(variantPatterns: Patterns): Record<string, VariantPolicyConfigObject> {
-    const variantPoliciesByPatterns: Record<string, VariantPolicyConfigObject> = {};
-    for (const [variantPattern, extensions] of Object.entries(variantPatterns)) {
-      if (extensions[DependencyResolverAspect.id]?.policy) {
-        variantPoliciesByPatterns[variantPattern] = extensions[DependencyResolverAspect.id]?.policy;
-      }
-    }
-    return variantPoliciesByPatterns;
-  }
-
-  private _updateVariantsPolicies(variantPatterns: Record<string, any>, updateVariantPolicies: string[]) {
-    for (const variantPattern of updateVariantPolicies) {
-      this.variants.setExtension(
-        variantPattern,
-        DependencyResolverAspect.id,
-        variantPatterns[variantPattern][DependencyResolverAspect.id],
-        { overrideExisting: true }
-      );
-    }
-    return this.dependencyResolver.persistConfig(this.path);
-  }
-
-  private async _saveManyComponentConfigFiles(componentConfigFiles: ComponentConfigFile[]) {
-    await Promise.all(
-      Array.from(componentConfigFiles).map(async (componentConfigFile) => {
-        await componentConfigFile.write({ override: true });
-      })
-    );
-  }
-
-  private async _installModules(options?: ModulesInstallOptions): Promise<ComponentMap<string>> {
-    this.logger.console(
-      `installing dependencies in workspace using ${chalk.cyan(this.dependencyResolver.getPackageManagerName())}`
-    );
-    this.logger.debug(`installing dependencies in workspace with options`, options);
-    const hasRootComponents = this.dependencyResolver.hasRootComponents();
-    // TODO: this make duplicate
-    // this.logger.consoleSuccess();
-    // TODO: add the links results to the output
-    await this.link({
-      linkTeambitBit: true,
-      legacyLink: true,
-      linkCoreAspects: false,
-      linkDepsResolvedFromEnv: !hasRootComponents,
-      linkNestedDepsInNM: false,
-    });
-    this.consumer.componentLoader.clearComponentsCache();
-    this.clearCache();
-    // TODO: pass get install options
-    const installer = this.dependencyResolver.getInstaller({});
-    const compDirMap = await this.getComponentsDirectory([]);
-    const mergedRootPolicy = this.dependencyResolver.getWorkspacePolicy();
-
-    const depsFilterFn = await this.generateFilterFnForDepsFromLocalRemote();
-
-    const pmInstallOptions: PackageManagerInstallOptions = {
-      dedupe: !hasRootComponents && options?.dedupe,
-      copyPeerToRuntimeOnRoot: options?.copyPeerToRuntimeOnRoot ?? true,
-      copyPeerToRuntimeOnComponents: options?.copyPeerToRuntimeOnComponents ?? false,
-      dependencyFilterFn: depsFilterFn,
-      overrides: this.dependencyResolver.config.overrides,
-      packageImportMethod: this.dependencyResolver.config.packageImportMethod,
-      rootComponents: hasRootComponents,
-    };
-    await installer.install(this.path, mergedRootPolicy, compDirMap, { installTeambitBit: false }, pmInstallOptions);
-    await this.link({
-      linkTeambitBit: false,
-      legacyLink: true,
-      linkCoreAspects: this.dependencyResolver.linkCoreAspects(),
-      linkDepsResolvedFromEnv: !hasRootComponents,
-      linkNestedDepsInNM: !this.isLegacy && !hasRootComponents,
-    });
-    await this.consumer.componentFsCache.deleteAllDependenciesDataCache();
-    return compDirMap;
-  }
-
-  async link(options: WorkspaceLinkOptions = {}): Promise<LinkResults> {
-    if (options.fetchObject) {
-      await this.importObjects();
-    }
-    options.consumer = this.consumer;
-    const compDirMap = await this.getComponentsDirectory([]);
-    const mergedRootPolicy = this.dependencyResolver.getWorkspacePolicy();
-    const linker = this.dependencyResolver.getLinker({
-      rootDir: this.path,
-      linkingOptions: options,
-    });
-    const res = await linker.link(this.path, mergedRootPolicy, compDirMap, options);
-    return res;
-  }
-
-  /**
-   * @param componentPath can be relative or absolute. supports Linux and Windows
-   */
-  async getComponentIdByPath(componentPath: PathOsBased): Promise<ComponentID | undefined> {
-    const relativePath = path.isAbsolute(componentPath) ? path.relative(this.path, componentPath) : componentPath;
-    const linuxPath = pathNormalizeToLinux(relativePath);
-    const bitId = this.consumer.bitMap.getComponentIdByPath(linuxPath);
-    if (bitId) {
-      return this.resolveComponentId(bitId);
-    }
-    return undefined;
-  }
-
-  /**
-   * Generate a filter to pass to the installer
-   * This will filter deps which are come from remotes which defined in scope.json
-   * those components comes from local remotes, usually doesn't have a package in a registry
-   * so no reason to try to install them (it will fail)
-   */
-  private async generateFilterFnForDepsFromLocalRemote() {
-    // TODO: once scope create a new API for this, replace it with the new one
-    const remotes = await this.scope._legacyRemotes();
-    const reg = await this.dependencyResolver.getRegistries();
-    const packageScopes = Object.keys(reg.scopes);
-    return (dependencyList: DependencyList): DependencyList => {
-      const filtered = dependencyList.filter((dep) => {
-        if (!(dep instanceof ComponentDependency)) {
-          return true;
-        }
-        if (remotes.isHub(dep.componentId.scope)) {
-          return true;
-        }
-        if (packageScopes.some((scope) => dep.packageName.startsWith(`@${scope}/`))) {
-          return true;
-        }
-        return false;
-      });
-      return filtered;
-    };
-  }
-
   /**
    * whether a scope is hosted by Bit cloud.
    * otherwise, it is self-hosted
@@ -2071,34 +1769,6 @@ your workspace.jsonc has this component-id set. you might want to remove/change 
         }
       }
 
-      throw err;
-    }
-  }
-
-  // TODO: replace with a proper import API on the workspace
-  private async importObjects() {
-    const importOptions: ImportOptions = {
-      ids: [],
-      verbose: false,
-      merge: false,
-      objectsOnly: true,
-      override: false,
-      writeDists: false,
-      writeConfig: false,
-      installNpmPackages: false,
-      importDependenciesDirectly: false,
-      importDependents: false,
-    };
-    const importer = new Importer(this, this.dependencyResolver);
-    try {
-      const res = await importer.import(importOptions, []);
-      return res;
-    } catch (err: any) {
-      // TODO: this is a hack since the legacy throw an error, we should provide a way to not throw this error from the legacy
-      if (err instanceof NothingToImport) {
-        // Do not write nothing to import warning
-        return undefined;
-      }
       throw err;
     }
   }
@@ -2298,17 +1968,6 @@ your workspace.jsonc has this component-id set. you might want to remove/change 
       }
     });
     return Promise.all(resolveMergedExtensionsP);
-  }
-
-  /**
-   * Uninstall the specified packages from dependencies.
-   *
-   * @param {string[]} the list of packages that should be removed from dependencies.
-   */
-  async uninstallDependencies(packages: string[]) {
-    this.dependencyResolver.removeFromRootPolicy(packages);
-    await this.dependencyResolver.persistConfig(this.path);
-    return this._installModules({ dedupe: true });
   }
 
   /**
